OBMIN-343 Changed name for entity and moved to another package
